feat(devpass): confirm plan switch + dashboard sidebar
- Add confirmation dialog before switching plans, with distinct
  upgrade vs downgrade messaging (no surprise charge on 'Switch to X').
- Downgrade backend: proration_behavior 'none' (no refund) and keep the
  paid allowance until renewal so downgrades never claw back credits.
- Split the DevPass dashboard into Usage/Billing/Settings routes behind a
  sidebar via a (main) route group; agents/feedback stay standalone.
- Billing page clarifies DevPass billing is separate from pay-as-you-go.
